+++ CHANGELOG.md
@@ -11,7 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-### Removed+
+### Removed
+
+## [1.3.5] - 2025-09-01
+
+### Fixed
+- AllaganLib.GameData
+  - FurnitureSource's now use the housing furniture sheet as a base, including the catalog information if applicable
 
 ## [1.3.4] - 2025-08-22